extract CacheDeprecatedReferences superclass from UpdatingDeprecatedReferences class to support other functionality such as detecting stale deprecations
